Soporte a ref. geografica del modelo segun BD
Referencia:
iMapa_info - 96313599a6751ab8b8036eb328163562ea9761df

+++ modelo/EntidadFederativa.h
@@ -9,8 +9,8 @@ class EntidadFederativa
 {
 public:
     QString nombre;
-    int latitud;
-    int longitud;
+    double latitud;
+    double longitud;
     double nHombresPorEntidad;
     double nMujeresPorEntidad;
